# backend/shadowvale_be/docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ShadowVale.API | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/analytics/versions/compare": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "parameters": [
          {
            "name": "A",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "B",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Source",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ContentVersionId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "MapCode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "From",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "To",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "SolverConfigurationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Family",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VersionComparisonDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionComparisonDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionComparisonDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/content/versions": {
      "get": {
        "tags": [
          "ContentVersions"
        ],
        "parameters": [
          {
            "name": "Status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfContentVersionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfContentVersionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfContentVersionDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ContentVersions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContentVersionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContentVersionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/content/versions/{id}": {
      "get": {
        "tags": [
          "ContentVersions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ContentVersions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContentVersionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContentVersionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ContentVersions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/content/versions/history": {
      "get": {
        "tags": [
          "ContentVersions"
        ],
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfPublicationHistoryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfPublicationHistoryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfPublicationHistoryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/content/versions/compare": {
      "get": {
        "tags": [
          "ContentVersions"
        ],
        "parameters": [
          {
            "name": "a",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "b",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ContentCompareDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentCompareDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentCompareDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/content/versions/{id}/validate": {
      "post": {
        "tags": [
          "ContentVersions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReportDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReportDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReportDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/content/versions/{id}/bundle": {
      "get": {
        "tags": [
          "ContentVersions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { }
            }
          }
        }
      }
    },
    "/api/content/versions/{id}/submit": {
      "post": {
        "tags": [
          "ContentVersions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/content/versions/{id}/approve": {
      "post": {
        "tags": [
          "ContentVersions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewNoteRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewNoteRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/content/versions/{id}/reject": {
      "post": {
        "tags": [
          "ContentVersions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectContentVersionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectContentVersionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RejectContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/content/versions/{id}/publish": {
      "post": {
        "tags": [
          "ContentVersions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishContentVersionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishContentVersionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PublishContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/content/versions/{id}/rollback": {
      "post": {
        "tags": [
          "ContentVersions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackContentVersionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackContentVersionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackContentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/game/content/versions/{versionId}/bundle": {
      "get": {
        "tags": [
          "Game"
        ],
        "parameters": [
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { }
            }
          }
        },
        "security": [
          {
            "GameKey": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContentCompareDto": {
        "required": [
          "a",
          "b",
          "hasChanges",
          "sections"
        ],
        "type": "object",
        "properties": {
          "a": {
            "$ref": "#/components/schemas/VersionRefDto"
          },
          "b": {
            "$ref": "#/components/schemas/VersionRefDto"
          },
          "hasChanges": {
            "type": "boolean"
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SectionDiffDto"
            }
          }
        }
      },
...
      "ContentVersionDto": {
        "required": [
          "id",
          "versionNo",
          "label",
          "changelog",
          "parentVersionId",
          "status",
          "revision",
          "schemaVersion",
          "isValidated",
          "validatedAt",
          "bundleChecksum",
          "validationErrors",
          "authoredBy",
          "submittedAt",
          "reviewedBy",
          "reviewedAt",
          "reviewNote",
          "publishedBy",
          "publishedAt",
          "archivedAt",
          "createdAt",
          "updatedAt",
          "counts"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "versionNo": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "label": {
            "type": "string"
          },
          "changelog": {
            "type": [
              "null",
              "string"
            ]
          },
          "parentVersionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "revision": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "schemaVersion": {
            "type": "string"
          },
          "isValidated": {
            "type": "boolean"
          },
          "validatedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "bundleChecksum": {
            "type": [
              "null",
              "string"
            ]
          },
          "validationErrors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ContentIssueDto"
            }
          },
          "authoredBy": {
            "$ref": "#/components/schemas/UserRefDto"
          },
          "submittedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "reviewedBy": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UserRefDto"
              }
            ]
          },
          "reviewedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "reviewNote": {
            "type": [
              "null",
              "string"
            ]
          },
          "publishedBy": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UserRefDto"
              }
            ]
          },
          "publishedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "archivedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "counts": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ContentCountsDto"
              }
            ]
          }
        }
      },
...
      "CreateContentVersionRequest": {
        "required": [
          "label"
        ],
        "type": "object",
        "properties": {
          "label": {
            "maxLength": 200,
            "type": "string"
          },
          "changelog": {
            "maxLength": 4000,
            "type": [
              "null",
              "string"
            ]
          },
          "baseVersionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
      "PagedResultOfContentVersionDto": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContentVersionDto"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "PagedResultOfPublicationHistoryDto": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicationHistoryDto"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "PublishContentVersionRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "maxLength": 500,
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RejectContentVersionRequest": {
        "required": [
          "note"
        ],
        "type": "object",
        "properties": {
          "note": {
            "maxLength": 2000,
            "type": "string"
          }
        }
      },
...
      "ReviewNoteRequest": {
        "type": "object",
        "properties": {
          "note": {
            "maxLength": 2000,
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RollbackContentVersionRequest": {
        "required": [
          "reason"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "maxLength": 500,
            "type": "string"
          }
        }
      },
...
      "UpdateContentVersionRequest": {
        "required": [
          "label"
        ],
        "type": "object",
        "properties": {
          "label": {
            "maxLength": 200,
            "type": "string"
          },
          "changelog": {
            "maxLength": 4000,
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ValidationReportDto": {
        "required": [
          "isValid",
          "issues",
          "bundleChecksum",
          "counts"
        ],
        "type": "object",
        "properties": {
          "isValid": {
            "type": "boolean"
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContentIssueDto"
            }
          },
          "bundleChecksum": {
            "type": [
              "null",
              "string"
            ]
          },
          "counts": {
            "$ref": "#/components/schemas/ContentCountsDto"
          }
        }
      },
      "VersionComparisonDto": {
        "required": [
          "a",
          "b"
        ],
        "type": "object",
        "properties": {
          "a": {
            "$ref": "#/components/schemas/VersionSnapshotDto"
          },
          "b": {
            "$ref": "#/components/schemas/VersionSnapshotDto"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Access token from POST /api/auth/login",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "GameKey": {
        "type": "apiKey",
        "description": "Key built into the game (config Game:ApiKeys)",
        "name": "X-Game-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}